becbilldesk_render_build-main/voice-agent/requirements.txt
# LiveKit Voice Agent Dependencies
livekit-agents>=0.12.0
livekit-plugins-openai>=0.7.1
livekit-plugins-deepgram>=0.5.0
livekit-plugins-cartesia>=1.2.8
livekit-plugins-silero>=0.6.10
python-dotenv>=1.0.0
aiohttp>=3.9.0
